requirements.txt
pygame~=2.5.2
Os-Maker
python-dotenv~=1.0.1